backend/requirements.txt
# FastAPI backend requirements for LexiBot

# Web framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# PDF processing
PyPDF2==3.0.1

# Google Gemini API
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
